refactor(agent-tools): extract script tools into agent-script-tools.ts
